requirements.txt
SwissArmyTransformer==0.4.12
omegaconf==2.3.0
pytorch_lightning==2.3.3
kornia==0.7.3
beartype==0.18.5
numpy==1.26.4
fsspec==2024.5.0
safetensors==0.4.3
imageio-ffmpeg==0.5.1
imageio==2.34.2
scipy==1.14.0
decord==0.6.0
wandb==0.17.5
deepspeed==0.14.4
rich==13.9.4
transformers==4.48.0
accelerate>=1.0.1
timm==1.0.11
transforms3d